Leverage Volar’s document drop edits provider
This moves logic away from the VSCode extension to the language server.
This means it can be implemented by other editors as well. Also this
makes it possible to provide smarter behaviour later, because the
language server has access to more MDX related context.
