+++ .changeset/ten-suns-hammer.md
@@ -1,0 +1,6 @@
+---
+"@mdx-js/language-service": patch
+"@mdx-js/language-server": patch
+---
+
+Implement Volar’s document drop edits
