+++ .changeset/twelve-cheetahs-occur.md
@@ -1,0 +1,5 @@
+---
+"vscode-mdx": minor
+---
+
+Use the language server to provide drop edits.
