+++ packages/language-service/lib/service-plugin.js
@@ -1,9 +1,30 @@
 /**
  * @typedef {import('@volar/language-service').ServicePlugin} ServicePlugin
+ * @typedef {import('@volar/language-service').DataTransferItem} DataTransferItem
  */
 
+import path from 'node:path/posix'
+import {toMarkdown} from 'mdast-util-to-markdown'
 import {fromPlace} from 'unist-util-lsp'
+import {URI, Utils} from 'vscode-uri'
 import {VirtualMdxFile} from './virtual-file.js'
+
+// https://github.com/microsoft/vscode/blob/1.83.1/extensions/markdown-language-features/src/languageFeatures/copyFiles/shared.ts#L29-L41
+const imageExtensions = new Set([
+  '.bmp',
+  '.gif',
+  '.ico',
+  '.jpe',
+  '.jpeg',
+  '.jpg',
+  '.png',
+  '.psd',
+  '.svg',
+  '.tga',
+  '.tif',
+  '.tiff',
+  '.webp'
+])
 
 /**
  * Create an Volar service plugin for MDX files.
@@ -18,6 +39,93 @@ export function createMdxServicePlugin() {
     name: 'mdx',
     create(context) {
       return {
+        async provideDocumentDropEdits(document, position, dataTransfer) {
+          const documentUri = URI.parse(document.uri)
+
+          /** @type {DataTransferItem | undefined} */
+          let textItem
+
+          /** @type {DataTransferItem | undefined} */
+          let uriListItem
+
+          for (const [mime, item] of dataTransfer) {
+            if (mime === 'text/plain') {
+              textItem = item
+              continue
+            }
+
+            if (mime === 'text/uri-list') {
+              uriListItem = item
+              continue
+            }
+
+            if (!mime.startsWith('image/')) {
+              continue
+            }
+
+            const file = item.asFile()
+            if (!file) {
+              continue
+            }
+
+            return {
+              insertText: toMarkdown({type: 'image', url: file.name}).trim(),
+              insertTextFormat: 1,
+              createDataTransferFile: [
+                {
+                  kind: 'create',
+                  uri: String(Utils.joinPath(documentUri, '..', file.name)),
+                  contentsMimeType: mime,
+                  options: {
+                    contents: file,
+                    ignoreIfExists: true
+                  }
+                }
+              ]
+            }
+          }
+
+          if (uriListItem) {
+            const value = await uriListItem.asString()
+            const uris = value.split(/\r?\n/)
+            /** @type {string[]} */
+            const content = []
+
+            for (const line of uris) {
+              try {
+                const uri = URI.parse(line, true)
+                const value =
+                  uri.scheme === documentUri.scheme
+                    ? path.relative(path.dirname(documentUri.path), uri.path)
+                    : line
+
+                content.push(
+                  toMarkdown(
+                    imageExtensions.has(path.extname(uri.path))
+                      ? {type: 'image', url: value}
+                      : {type: 'text', value}
+                  ).trim()
+                )
+              } catch {
+                continue
+              }
+            }
+
+            return {
+              insertText: content.join(' '),
+              insertTextFormat: 1
+            }
+          }
+
+          if (textItem) {
+            const string = await textItem.asString()
+            return {
+              insertText: string,
+              insertTextFormat: 1
+            }
+          }
+        },
+
         provideSemanticDiagnostics(document) {
           const file = getVirtualMdxFile(document.uri)
