+++ packages/vscode-mdx/package.json
@@ -44,14 +44,11 @@
     "vscode:prepublish": "npm run build"
   },
   "devDependencies": {
-    "@types/mdast": "^4.0.0",
-    "@types/node": "^20.0.0",
     "@types/vscode": "^1.82.0",
     "@volar/language-server": "2.0.0-alpha.13",
     "@volar/vscode": "2.0.0-alpha.13",
     "@vscode/vsce": "^2.0.0",
     "esbuild": "^0.19.0",
-    "mdast-util-to-markdown": "^2.0.0",
     "ovsx": "^0.8.0",
     "undici": "^5.0.0"
   },
